requirements.txt
# Chainlit for chat interface
chainlit>=1.0.0

# Azure OpenAI for AI
openai>=1.10.0

# Pydantic for structured outputs and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# FastAPI for REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# MongoDB driver
motor>=3.3.0
pymongo>=4.6.0

# Environment management
python-dotenv>=1.0.0

# HTTP client for tool calling
httpx>=0.26.0

# Additional utilities
python-multipart>=0.0.6
